mod install-nvim.sh and links.sh

+++ install.d/link.sh
@@ -2,6 +2,28 @@
 #!/bin/bash
 DOTDIR=$HOME/dotfiles
 
-ln -s $DOTDIR/.tmux.conf $HOME
-ln -s $DOTDIR/.config $HOME
-ln -s $DOTDIR/.bash_profile $HOME
+TMUXDIR=$HOME/.tmux.conf
+if [ -e $TMUXDIR ]; then
+	rm -r $TMUXDIR
+	ln -s $DOTDIR/.tmux.conf $HOME
+else
+	ln -s $DOTDIR/.tmux.conf $HOME
+fi
+
+
+CONFDIR=$HOME/.config
+if [ -e $CONFDIR ]; then
+	rm -r $CONFDIR
+	ln -s $DOTDIR/.config $HOME
+else
+	ln -s $DOTDIR/.config $HOME
+fi
+
+
+BASHPROFILEDIR=$HOME/.bash_profile
+if [ -e $BASHPROFILEDIR ]; then
+	rm -r $BASHPROFILEDIR
+	ln -s $DOTDIR/.bash_profile $HOME
+else
+	ln -s $DOTDIR/.bash_profile $HOME
+fi
